Add Alerts, Queries and Mutations Hooks to Stake Component
Updates were made to the Stake component to include new hooks for staking queries and mutations. Additionally, an alert component was integrated to handle success and error messaging. This simplifies the logic in the staking component by shifting responsibilities to the newly added hooks, resulting in cleaner and more maintainable code.