Mainnet only deployment
TODO: add testnet support

Refactor AlertComponent and modify chain configurations

Changed the AlertTitle to AlertDescription in the AlertComponent. The chain configurations in the providers.tsx file were also modified. More specifically, made adjustments related to the chainId and its details, changes to the chains array, and modified the transports. In the useStakingMutations hook, the code was modified to directly fetch the seed token address and staking contract address from environment variables.